Log closest bookmark on annotationChange
As a proof of concept, add logic to print the closest `bookmark` parent
whenever an annotation change occurs (i.e. added or edited)

Refactor `setBookmarkCoordinates` to ensure that the most up to date
`bookmarkCoordinates` can be accessed within the `annotationChanged`
event listener. See

https://stackoverflow.com/questions/55265255/react-usestate-hook-event-handler-using-initial-state

For more details on why the refactor for `setBookmarkCoordinates` was
necessary